- fixed baseEnemy takeDamage function - ult ability on thea works - changed prefab spawn position - thea can damage and kill the enemy

+++ Raid of Reason/Assets/Scenes/Test Scenes/SceneDenver/Scripts/BaseEnemy.cs
@@ -10,10 +10,10 @@ public abstract class BaseEnemy : MonoBehaviour
     [SerializeField] protected float m_viewRange;
 
 	[Tooltip("Enemy's maximum amount of health.")]
-	[SerializeField] protected float m_maxHealth;
+	[SerializeField] protected int m_maxHealth;
 
-	protected float m_health;
-	public float Health { get { return m_health; } }
+	protected int m_health;
+	public int Health { get { return m_health; } }
 
     protected NavMeshAgent m_navMeshAgent;
     protected GameObject[] m_players;
